Working on Implement SFU SSO
For now, when user press login button on the front-end
deployment, it will redirect user to SFU SSO.

When the user provide credentials there, it will give ticket to the back-end deployment. The back-end will use that ticket to ask SFU SSO again and SFU CAS will return a XML file about the validation.

In that XML file, the SFU user id is contained. The back-end will use that user id to search our user repository. If the user does not exist, it will create new user. If the user exist, it it will search the user Long type id.

I'm working on session to connect the validation between front-end and back-end.